+++ R/newCorrespondenceTable.R
@@ -3,10 +3,6 @@
 #' @param Tables A string of type character containing the name of a csv file which contains the names of the files that
 #'   contain the classifications and the intermediate correspondence tables OR a list of vectors with the names of the dataframes
 #'   classifications and the intermediate correspondence tables (see "Details" below).
-#' @param CSVout The preferred name for the \emph{output csv files} that will contain the candidate correspondence table
-#'   and information about the classifications involved. The valid values are \code{NULL} or strings of type \code{character}.
-#'   If the selected value is \code{NULL}, the default, no output file is produced. If the value is a string, then the output
-#'   is exported into two csv files whose names contain the provided name (see "Value" below).
 #' @param Reference The reference classification among A and B. If a classification is the reference to the other, and hence
 #'  \emph{hierarchically superior} to it, each code of the other classification is expected to be mapped to at most one code
 #'   of the reference classification. The valid values are \code{"none"}, \code{"A"}, and \code{"B"}. If the selected value
@@ -16,9 +12,13 @@
 #'   no code for classification A or no code for classification B. The default value is \code{0.2}. The valid values are
 #'   real numbers in the interval [0, 1].
 #' @param Redundancy_trim An argument in the function containing the logical values \code{TRUE} or \code{FALSE}
-#' used to facilitate the trimming of the redundant records.
-#' The default value is \code{TRUE}, which removes all redundant records.
-#' The other values is \code{FALSE}, which shows redundant records together with the redundancy flag.
+#'   used to facilitate the trimming of the redundant records.
+#'   The default value is \code{TRUE}, which removes all redundant records.
+#'   The other values is \code{FALSE}, which shows redundant records together with the redundancy flag.
+#' @param CSVout The preferred name for the \emph{output csv files} that will contain the candidate correspondence table
+#'   and information about the classifications involved. The valid values are \code{NULL} or strings of type \code{character}.
+#'   If the selected value is \code{NULL}, the default, no output file is produced. If the value is a string, then the output
+#'   is exported into two csv files whose names contain the provided name (see "Value" below).
 #' @export
 #' @details
 #' File and file name requirements:
@@ -169,11 +169,13 @@
 #'                  col.names = FALSE,
 #'                  sep = ",")
 #'
-#'      NCT<-newCorrespondenceTable(file.path(tmp_dir,"example.csv"),
-#'                                  file.path(tmp_dir,"newCorrespondenceTable.csv"),
-#'                                  "A",
-#'                                  0.56,
-#'                                  FALSE)
+#'       NCT <- newCorrespondenceTable(
+#'         Tables            = file.path(tmp_dir, "example.csv"),
+#'         Reference         = "A",
+#'         MismatchTolerance = 0.56,
+#'         Redundancy_trim   = FALSE,
+#'         CSVout= file.path(tmp_dir, "newCorrespondenceTable.csv")
+#'         )
 #'
 #'      summary(NCT)
 #'      head(NCT$newCorrespondenceTable)
@@ -182,8 +184,13 @@
 #'      unlink(csv_files)
 #'     }
 
-newCorrespondenceTable <- function(Tables, CSVout = NULL, Reference = "none", MismatchTolerance = 0.2, Redundancy_trim = TRUE) {
-
+newCorrespondenceTable <- function(
+    Tables,
+    Reference = "none",
+    MismatchTolerance = 0.2,
+    Redundancy_trim = TRUE,
+    CSVout = NULL
+){
   # Check if the Tables is a File or a list of vectors
   # Check if the file that contains the names of both classifications and correspondence tables exists in working directory